Removing reconnectRetries from all plugins. Removing deserializing splunk json config in tests.
