+++ contrib/storage-druid/src/main/java/org/apache/drill/exec/store/druid/DruidStoragePluginConfig.java
@@ -42,7 +42,6 @@ public class DruidStoragePluginConfig extends StoragePluginConfig {
     @JsonProperty("brokerAddress") String brokerAddress,
     @JsonProperty("coordinatorAddress") String coordinatorAddress,
     @JsonProperty("averageRowSizeBytes") Integer averageRowSizeBytes) {
-    super(DISABLED_RECONNECT_RETRIES);
     this.brokerAddress = brokerAddress;
     this.coordinatorAddress = coordinatorAddress;
     this.averageRowSizeBytes =
