+++ contrib/storage-hbase/src/main/java/org/apache/drill/exec/store/hbase/HBaseStoragePluginConfig.java
@@ -50,7 +50,6 @@ public class HBaseStoragePluginConfig extends StoragePluginConfig implements Dri
 
   @JsonCreator
   public HBaseStoragePluginConfig(@JsonProperty("config") Map<String, String> props, @JsonProperty("size.calculator.enabled") Boolean sizeCalculatorEnabled) {
-    super(DISABLED_RECONNECT_RETRIES);
     this.config = props;
     if (config == null) {
       config = Maps.newHashMap();
