+++ contrib/storage-kafka/src/main/java/org/apache/drill/exec/store/kafka/KafkaStoragePluginConfig.java
@@ -40,7 +40,6 @@ public class KafkaStoragePluginConfig extends StoragePluginConfig {
 
   @JsonCreator
   public KafkaStoragePluginConfig(@JsonProperty("kafkaConsumerProps") Map<String, String> kafkaConsumerProps) {
-    super(DISABLED_RECONNECT_RETRIES);
     this.kafkaConsumerProps = new Properties();
     this.kafkaConsumerProps.putAll(kafkaConsumerProps);
     logger.debug("Kafka Consumer Props {}", this.kafkaConsumerProps);
