+++ contrib/storage-kudu/src/main/java/org/apache/drill/exec/store/kudu/KuduStoragePluginConfig.java
@@ -32,7 +32,6 @@ public class KuduStoragePluginConfig extends StoragePluginConfig {
 
   @JsonCreator
   public KuduStoragePluginConfig(@JsonProperty("masterAddresses") String masterAddresses) {
-    super(DISABLED_RECONNECT_RETRIES);
     this.masterAddresses = masterAddresses;
   }
