+++ contrib/storage-splunk/src/main/resources/bootstrap-storage-plugins.json
@@ -2,13 +2,13 @@
   "storage":{
     "splunk" : {
       "type":"splunk",
+      "enabled": false,
       "username": "admin",
       "password": "changeme",
       "hostname": "localhost",
       "port": 8089,
       "earliestTime": "-14d",
       "latestTime": "now",
-      "enabled": false,
       "reconnectRetries": 2
     }
   }
