+++ contrib/storage-splunk/src/test/java/org/apache/drill/exec/store/splunk/SplunkConnectionTest.java
@@ -49,7 +49,8 @@ public class SplunkConnectionTest extends SplunkBaseTest {
               SPLUNK_STORAGE_PLUGIN_CONFIG.getPort(),
               SPLUNK_STORAGE_PLUGIN_CONFIG.getEarliestTime(),
               SPLUNK_STORAGE_PLUGIN_CONFIG.getLatestTime(),
-              null
+              null,
+              SPLUNK_STORAGE_PLUGIN_CONFIG.getReconnectRetries()
       );
       SplunkConnection sc = new SplunkConnection(invalidSplunkConfig);
       sc.connect();
