+++ contrib/storage-splunk/src/test/java/org/apache/drill/exec/store/splunk/SplunkPluginTest.java
@@ -307,11 +307,13 @@ public class SplunkPluginTest extends SplunkBaseTest {
       loginArgs.setUsername(SPLUNK_STORAGE_PLUGIN_CONFIG.getUsername());
       splunk.when(() -> Service.connect(loginArgs))
           .thenThrow(new RuntimeException("Fail first connection to Splunk"))
-          .thenReturn(new Service(loginArgs));
-      new SplunkConnection(SPLUNK_STORAGE_PLUGIN_CONFIG); //it fails, in case "reconnectRetries": 1 is specified in configs
+          .thenThrow(new RuntimeException("Fail second connection to Splunk"))
+          .thenThrow(new RuntimeException("Fail third connection to Splunk"))
+          .thenReturn(new Service(loginArgs)); // fourth connection is successful
+      new SplunkConnection(SPLUNK_STORAGE_PLUGIN_CONFIG); // it will fail, in case "reconnectRetries": 1 is specified in configs
       splunk.verify(
           () -> Service.connect(loginArgs),
-          times(2)
+          times(4)
       );
     }
   }
