--- contrib/storage-splunk/src/test/resources/bootstrap-storage-plugins.json
@@ -1,15 +1,0 @@
-{
-  "storage":{
-    "splunk" : {
-      "type":"splunk",
-      "enabled": false,
-      "username": "admin",
-      "password": "password",
-      "hostname": "localhost",
-      "port": 8089,
-      "earliestTime": "1",
-      "latestTime": "now",
-      "reconnectRetries": 2
-    }
-  }
-}
