+++ exec/java-exec/src/main/java/org/apache/drill/exec/store/mock/MockStorageEngineConfig.java
@@ -32,7 +32,6 @@ public class MockStorageEngineConfig extends StoragePluginConfig {
 
   @JsonCreator
   public MockStorageEngineConfig(@JsonProperty("url") String url) {
-    super(DISABLED_RECONNECT_RETRIES);
     this.url = url;
   }
