+++ exec/java-exec/src/main/java/org/apache/drill/exec/store/sys/SystemTablePluginConfig.java
@@ -28,9 +28,7 @@ public class SystemTablePluginConfig extends StoragePluginConfig {
 
   public static final SystemTablePluginConfig INSTANCE = new SystemTablePluginConfig();
 
-  private SystemTablePluginConfig() {
-    super(DISABLED_RECONNECT_RETRIES);
-  }
+  private SystemTablePluginConfig() { }
 
   @Override
   public boolean equals(Object o) {
