+++ exec/java-exec/src/test/java/org/apache/drill/exec/store/BasePluginRegistryTest.java
@@ -88,7 +88,6 @@ public class BasePluginRegistryTest extends BaseTest {
 
     @JsonCreator
     public StoragePluginFixtureConfig(@JsonProperty("mode") String mode) {
-      super(DISABLED_RECONNECT_RETRIES);
       this.mode = mode;
     }
